@@ -1,11 +1,10 @@
 name: nine_grid_view
 description: Flutter NineGridView & DragSortView, Similar to Weibo / WeChat nine grid view controls to display pictures.
-version: 1.0.7
-author: Sky24n <sky24no@gmail.com>
+version: 2.0.0
 homepage: https://github.com/flutterchina/nine_grid_view
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: ">=2.12.0-259.9.beta <3.0.0"
 
 dependencies:
   flutter:
