@@ -10,7 +10,7 @@ class SinglePicturePage extends StatefulWidget {
 }
 
 class _SinglePicturePageState extends State<SinglePicturePage> {
-  List<ImageBean> imageList = List();
+  List<ImageBean> imageList = [];
 
   @override
   void initState() {
@@ -22,13 +22,13 @@ class _SinglePicturePageState extends State<SinglePicturePage> {
     int itemCount = 1;
 
     ImageBean imageBean = imageList[_index];
-    String url = imageBean.middlePath;
+    String url = imageBean.middlePath!;
 
-    int originalWidth;
-    int originalHeight;
+    int? originalWidth;
+    int? originalHeight;
 
-    Image bigImage;
-    String bigImageUrl;
+    Image? bigImage;
+    String? bigImageUrl;
 
     // Single picture display big picture.
     if (itemCount == 1) {
